requirements.txt
Click==7.0
DateTime==4.3
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.2
pandas==0.23.4
pymssql==2.1.4
python-crontab==2.3.5
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
SQLAlchemy==1.2.12
Werkzeug==0.14.1
zope.interface==4.5.0
